+++ src/DependencyInjection/CompilerPass/RegisterTwigUndefinedCallablePass.php
@@ -1,0 +1,28 @@
+<?php declare(strict_types=1);
+
+namespace BabDev\PagerfantaBundle\DependencyInjection\CompilerPass;
+
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Reference;
+use Twig\Environment;
+
+/**
+ * @internal
+ */
+final class RegisterTwigUndefinedCallablePass implements CompilerPassInterface
+{
+    public function process(ContainerBuilder $container): void
+    {
+        if (!$container->hasDefinition('twig')) {
+            return;
+        }
+
+        if (!method_exists(Environment::class, 'registerUndefinedFunctionCallback')) {
+            return;
+        }
+
+        $container->getDefinition('twig')
+            ->addMethodCall('registerUndefinedFunctionCallback', [[new Reference('pagerfanta.undefined_callable_handler'), 'onUndefinedFunction']]);
+    }
+}